requirements.txt
google-cloud-vision
PyYaml
Flask
Flask_sqlalchemy